Add stage_execution header
