+++ node/silkworm/stagedsync/sync_loop.cpp
@@ -20,6 +20,7 @@
 
 #include <silkworm/stagedsync/stage_blockhashes.hpp>
 #include <silkworm/stagedsync/stage_senders.hpp>
+#include <silkworm/stagedsync/stage_execution.hpp>
 #include <silkworm/stagedsync/stagedsync.hpp>
 
 namespace silkworm::stagedsync {
